# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZK-DPP Royalty Protocol API
  description: |
    API documentation for the Zero-Knowledge Digital Product Passport (DPP) Royalty Protocol.

    This protocol enables privacy-preserving compliance verification for supply chain data,
    allowing suppliers to prove compliance without exposing trade secrets.

    ## Services Overview

    The system consists of three microservices:

    | Service | Port | Purpose |
    |---------|------|---------|
    | verify-gateway | 3001 | ZK proof verification and receipt issuance |
    | dpp-builder | 3002 | Product management and DPP view composition |
    | metering-billing | 3003 | Usage tracking and settlement management |

    ## Authentication

    Most endpoints require JWT authentication via Keycloak. Include the token in the Authorization header:

    ```
    Authorization: Bearer <token>
    ```

    ## Access Levels

    DPP views are tiered by access level:
    - **PUBLIC**: No authentication required
    - **LEGIT_INTEREST**: Requires `auditor` or `authority` role
    - **AUTHORITY**: Requires `authority` role
  version: 0.1.0
  license:
    name: Apache-2.0 OR MIT
    url: https://github.com/hadijannat/zk-dpp-royalty-protocol/blob/main/LICENSE
  contact:
    name: ZK-DPP Team
    url: https://github.com/hadijannat/zk-dpp-royalty-protocol
servers:
  - url: http://localhost:3001
    description: Verify Gateway (local development)
  - url: http://localhost:3002
...
  - url: http://localhost:3003
...
paths:
...
  /settlements/{id}/submit-on-chain:
    post:
      tags: [Blockchain]
      summary: Submit settlement to blockchain
      description: |
        Submits a finalized settlement statement to the blockchain smart contract.
        Requires blockchain configuration to be enabled.
      operationId: submitSettlementOnChain
      servers:
        - url: http://localhost:3003
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [supplierWallet]
              properties:
                supplierWallet:
                  type: string
                  pattern: '^0x[a-fA-F0-9]{40}$'
                  description: Ethereum wallet address for the supplier
            example:
              supplierWallet: "0x1234567890123456789012345678901234567890"
      responses:
        '200':
          description: Settlement submitted to blockchain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainTransactionResponse'
        '400':
          description: Invalid request or statement not finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Statement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /settlements/{id}/finalize-on-chain:
    post:
      tags: [Blockchain]
      summary: Finalize settlement on blockchain
      description: |
        Finalizes a settlement on the blockchain after the dispute window has passed.
        This transfers funds to the supplier's claimable balance.
      operationId: finalizeSettlementOnChain
      servers:
        - url: http://localhost:3003
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Settlement finalized on blockchain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainTransactionResponse'
        '400':
          description: Dispute window not passed or invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /settlements/{id}/blockchain-status:
    get:
      tags: [Blockchain]
      summary: Get blockchain status
      description: Returns the blockchain status of a settlement
      operationId: getSettlementBlockchainStatus
      servers:
        - url: http://localhost:3003
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Blockchain status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainStatusResponse'
        '404':
          description: Statement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /supplier-wallets:
    post:
      tags: [Blockchain]
      summary: Register supplier wallet
      description: Registers an Ethereum wallet address for a supplier
      operationId: registerSupplierWallet
      servers:
        - url: http://localhost:3003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [supplierId, walletAddress]
              properties:
                supplierId:
                  type: string
                walletAddress:
                  type: string
                  pattern: '^0x[a-fA-F0-9]{40}$'
      responses:
        '201':
          description: Wallet registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  wallet:
                    $ref: '#/components/schemas/SupplierWallet'
        '400':
          description: Invalid wallet address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /supplier-wallets/{supplierId}:
    get:
      tags: [Blockchain]
      summary: Get supplier wallet
      description: Returns the registered wallet for a supplier
      operationId: getSupplierWallet
      servers:
        - url: http://localhost:3003
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Supplier wallet
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallet:
                    $ref: '#/components/schemas/SupplierWallet'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Keycloak
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        success:
          type: boolean
          default: false
...
    BlockchainTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
        txHash:
          type: string
          description: Transaction hash on the blockchain
        blockNumber:
          type: integer
          description: Block number where transaction was included
        error:
          type: string
    BlockchainStatusResponse:
      type: object
      properties:
        statementId:
          type: string
        blockchainStatus:
          type: string
          enum: [NOT_SUBMITTED, PENDING, SUBMITTED, FINALIZED, FAILED]
        txHash:
          type: string
        blockNumber:
          type: integer
        chainSubmittedAt:
          type: string
          format: date-time
        chainFinalizedAt:
          type: string
          format: date-time
        isFinalizable:
          type: boolean
          description: Whether the dispute window has passed
        remainingDisputeTime:
          type: integer
          description: Seconds remaining in dispute window
    SupplierWallet:
      type: object
      properties:
        supplier_id:
          type: string
        wallet_address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        created_at:
          type: string
          format: date-time